requirements.txt
Flask
Pillow
tensorflow
keras
urllib3
gunicorn